16-Bit FSR Cleanup & Fixes
This change removes the 16-bit option in the FSRCommon.hlsl in favor of
an automatic solution based on REAL_IS_HALF. Users of the FSR code no
longer need to explicitly opt-in to the 16-bit implementation. It will
now be enabled automatically whenever the target platform has support.

There are currently two exceptions for the logic that automatically
enables the 16-bit FSR implementation:

1. Some vendors have issues with 16-bit floating point casts to uint16
   which are required for FSR's math approximations. This issue only
   occurs on DX11 drivers, so the 16-bit implementation is currently
   disabled whenever DX11 is in use until this is fixed.
2. There's currently a known shader compiler issue that affects
   platforms running Metal with FP16 support. (Case 1387697) Due to
   this issue, the 16-bit implementation is also disabled for Metal.

This change also contains a small modification for the uint to float
conversion logic used inside FSRUtils.cs:SetRcasConstants. The updated
logic avoids an unnecessary memory allocation which was causing
failures in URP's graphics tests.